Refactor autopilot and add basic tests
